Replace Test-ModuleManifest with Import-PowerShellDataFile
Test-ModuleManifest will return the version in them module, but also tests that all the files and dependencies are in place. There's no reason to install dependencies on the publishing agent, when all we want is the version number. Import-PowerShellDataFile will read any generic psd1 file, and so we can use it to read the 'ModuleVersion' property directly of the PSD1.